Made letsencrypt-enable.sh to be non interactive

+++ letsencrypt-enable.sh
@@ -10,29 +10,36 @@ DOMAINS=(pegbtech-demo.tomspirit.me www.pegbtech-demo.tomspirit.me)
 RSA_KEY_SIZE=4096
 DATA_PATH="./pegb_web/pegb-certbot"
 EMAIL="tome.petkovski986@gmail.com" # Adding a valid address is strongly recommended
-STAGING=0 # Set to 1 if you're testing your setup to avoid hitting request limits
+STAGING=1 # Set to 1 if you're testing your setup to avoid hitting request limits
 
-# Nginx and certbot service names from docker-compose
+## Nginx and certbot service names from docker-compose
 NGINX_SERVICE=pegb-proxy 
 CERTBOT_SERVICE=pegb-certbot
 
 
-if [ -d "$DATA_PATH" ]; then
-  read -p "Existing certificate domains found for: $DOMAINS. Continue and replace existing certificates? (y/N) " decision
-  if [ "$decision" != "Y" ] && [ "$decision" != "y" ]; then
-    exit
-  fi
+## Ask the user to confirm if he wishes to delete the current certificates in $DATA_PATH
+if [ -d "$DATA_PATH" ]
+then
+	read -p "Existing certificate domains found for: $DOMAINS. Continue and replace existing certificates? (y/N) " decision
+	
+	if [ "$decision" != "Y" ] && [ "$decision" != "y" ]
+	then
+		exit
+	fi
 fi
 
 
-if [ ! -e "$DATA_PATH/conf/options-ssl-nginx.conf" ] || [ ! -e "$DATA_PATH/conf/ssl-dhparams.pem" ]; then
-  echo "### Downloading recommended TLS parameters ..."
-  mkdir -p "$DATA_PATH/conf"
-  curl -s https://raw.githubusercontent.com/certbot/certbot/master/certbot-nginx/certbot_nginx/tls_configs/options-ssl-nginx.conf > "$DATA_PATH/conf/options-ssl-nginx.conf"
-  curl -s https://raw.githubusercontent.com/certbot/certbot/master/certbot/ssl-dhparams.pem > "$DATA_PATH/conf/ssl-dhparams.pem"
-  echo
+## Download reccomended TLS parameters as per the Let's Encrypt reccomendations
+if [ ! -e "$DATA_PATH/conf/options-ssl-nginx.conf" ] || [ ! -e "$DATA_PATH/conf/ssl-dhparams.pem" ]
+then
+	echo "### Downloading recommended TLS parameters ..."
+	mkdir -p "$DATA_PATH/conf"
+	curl -s https://raw.githubusercontent.com/certbot/certbot/master/certbot-nginx/certbot_nginx/tls_configs/options-ssl-nginx.conf > "$DATA_PATH/conf/options-ssl-nginx.conf"
+	curl -s https://raw.githubusercontent.com/certbot/certbot/master/certbot/ssl-dhparams.pem > "$DATA_PATH/conf/ssl-dhparams.pem"
+	echo
 fi
 
+## Creating dummy certificate to allow nginx to start
 echo "### Creating dummy certificate for $DOMAINS ..."
 KEY_PATH="/etc/letsencrypt/live/$DOMAINS"
 /usr/bin/mkdir -p "$DATA_PATH/conf/live/$DOMAINS"
@@ -48,21 +55,29 @@ echo "### Starting nginx ..."
 /usr/local/bin/docker-compose up --force-recreate -d $NGINX_SERVICE
 echo
 
+## If staging is set to other than 0 the dummy certificates created previously will not be deleted
+if [[ $STAGING -eq 0 ]]
+then
+	echo "### Deleting dummy certificate for $DOMAINS ..."
+	/usr/local/bin/docker-compose run --rm --entrypoint "\
+  	rm -Rf /etc/letsencrypt/live/$DOMAINS && \
+  	rm -Rf /etc/letsencrypt/archive/$DOMAINS && \
+  	rm -Rf /etc/letsencrypt/renewal/$DOMAINS.conf" $CERTBOT_SERVICE
+	echo
+else
+	echo "### STAGING mote set, moving on ..."
+fi
 
-echo "### Deleting dummy certificate for $DOMAINS ..."
-/usr/local/bin/docker-compose run --rm --entrypoint "\
-  rm -Rf /etc/letsencrypt/live/$DOMAINS && \
-  rm -Rf /etc/letsencrypt/archive/$DOMAINS && \
-  rm -Rf /etc/letsencrypt/renewal/$DOMAINS.conf" $CERTBOT_SERVICE
-echo
 
-
+## Request the certificate
 echo "### Requesting Let's Encrypt certificate for $DOMAINS ..."
 #Join $DOMAINS to -d args
 domain_args=""
-for domain in "${DOMAINS[@]}"; do
-  domain_args="$domain_args -d $domain"
+for domain in "${DOMAINS[@]}"
+do
+	domain_args="$domain_args -d $domain"
 done
+
 
 # Select appropriate email arg
 case "$EMAIL" in
@@ -78,6 +93,7 @@ if [ $STAGING != "0" ]; then staging_arg="--staging"; fi
     $staging_arg \
     $email_arg \
     $domain_args \
+    --non-interactive \
     --rsa-key-size $RSA_KEY_SIZE \
     --agree-tos \
     --force-renewal" $CERTBOT_SERVICE
